Refactor rrdstep.py
Fixes some of my laziness from when this was just a hacky script that I put together and abandoned after it got smoked by the Rust rewrite (that's fixed now).
Should be identical to how it was before, with the exception that it now supports increasing step and the return codes for failure have changed.